refactor: extract BroadcastTab into broadcast/ sub-components (7.4)
Split the 507-line monolithic BroadcastTab into focused, testable units:
  broadcast/broadcastTypes.ts   — shared SectorClass / SectorAnalysisData types
  broadcast/broadcastUtils.ts   — classifySectorEntries + buildSectorAnalysis
  broadcast/TimingTower.tsx     — timing tower panel
  broadcast/SectorAnalysis.tsx  — sector analysis table + legend
  broadcast/SpeedTrap.tsx       — speed-trap measurement table
  broadcast/DriverCards.tsx     — broadcast-style driver cards grid
BroadcastTab.tsx is now ~100 lines that compose the four sub-components.
